MOD/RE-FACTOR/ Gran modularización del código
