+++ src/routes/comment/comment-Routes.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import db from "../../../models/index";
-import { validateNewComment } from "../../auxiliary/CommentValidators";
+import { validateNewComment } from "../../validators/CommentValidators";
 import { IComment, ICommentResponse } from "../../types/commentTypes";
 import { IImage } from "../../types/imageTypes";
